Adapt filtering logic of anonymous functions to typescript language server
There was a filtering logic in place for anonymous functions before, which
filtered out the ones with a '<function>' description. This check had to be
extended, since in TypeScript 3.2.x the description for anonymous functions
were changed so that they are not always '<function>', but in case of
CallExpressions they also contain the method name, argument list and a
'callback' suffix.

This fixes #34